enabled floating-point slide-show times. closes #2

+++ src/ch/awae/imgtagger/AutoPlayController.scala
@@ -1,13 +1,13 @@
 package ch.awae.imgtagger
 
-class AutoPlayController(val delay: Int, val random: Boolean, val manager: WindowManager) extends Thread {
-  
+class AutoPlayController(val delay: Double, val random: Boolean, val manager: WindowManager) extends Thread {
+
   this.start()
-  
+
   override def run {
     while (!this.isInterrupted())
       try {
-        Thread.sleep(delay * 1000L)
+        Thread.sleep((delay * 1000L).toLong)
         if (random)
           manager.navRandom
         else
